Use `GoToPointCommand` in `AutoAlignCommand`, separate "align with nearest" logic
