+++ src/main/java/frc/robot/commands/dropping/AutoAlignGrid.java
@@ -8,7 +8,7 @@ import frc.robot.subsystems.drivetrain.BaseSwerveSubsystem;
 import frc.robot.subsystems.tiltedelevator.TiltedElevatorSubsystem;
 
 public class AutoAlignGrid {
-    private final AutoAlignCommand alignToClosestCommand;
+    private final AutoAlignToClosestCommand alignToClosestCommand;
     private final AutoAlignCommand[] setTargetCommands;
 
     private final ShuffleboardTab shuffleboardTab = Shuffleboard.getTab("Grid");
@@ -25,7 +25,7 @@ public class AutoAlignGrid {
         BaseSwerveSubsystem swerveSubsystem, TiltedElevatorSubsystem tiltedElevatorSubsystem,
         boolean isRed // TODO: better way of passing this
     ) {
-        alignToClosestCommand = new AutoAlignCommand(swerveSubsystem, tiltedElevatorSubsystem, isRed);
+        alignToClosestCommand = new AutoAlignToClosestCommand(swerveSubsystem, tiltedElevatorSubsystem, isRed);
 
         PlacePosition[] positions = PlacePosition.values();
         setTargetCommands = new AutoAlignCommand[positions.length];
@@ -48,7 +48,7 @@ public class AutoAlignGrid {
      * Gets the command to align to the closest node.
      * @return The command to align to the closest node.
      */
-    public AutoAlignCommand getAlignToClosestCommand() {
+    public AutoAlignToClosestCommand getAlignToClosestCommand() {
         return alignToClosestCommand;
     }
